Make apollo-server-express example complete
It wasn't obvious to me that I was meant to call the method, and even then, after getting the error about awaiting and startup issues, it wasn't (and still is not) obvious to me when I would call `await startApolloServer` or with no await.  I suppose I'm in a synchronous main method so await would not have any meaning, but that's kind of in the weeds for me and this seems like a beginner should be able cut and paste.